Add VPN server IP to allowed addresses in TrafPol
During a user's connection attempt, the XML profile can get updated and
the TrafPol component restarts with the new settings of the XML profile.
This can cause the VPN connection to fail when the user's connection
request arrives in OC-Daemon and TrafPol has not resolved the IP address
of the VPN server in the allowed hosts yet. So, add the IP address of
the VPN server in the user's connection request to the allowed addresses
in TrafPol to make sure access to the VPN server is allowed even when
TrafPol is restarting.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>