Parse Login Info from whole lines in Client
The Client captures stdout into a buffer, splits it into lines with
strings.Fields() and parses the Login Info in those lines. But Login
Info values like the Cookie can contain white space characters which
results in parsing incomplete lines. So, extract whole lines from the
buffer with a bufio.Scanner and use them to parse the Login Info.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>