Add support for loading command lists from file
Add support for loading command lists from a file. The file can contain
a subset or all command lists. Only command lists in the file are
changed. The remaining command lists keep their default values. The
default path of the file is /var/lib/oc-daemon/command-lists.json and
can be changed in the OC-Daemon configuration file.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>