Add ocserv tests
Add tests for OC-Daemon that use ocserv as VPN server. The tests use
podman-compose to run OC-Daemon, ocserv and two web servers as
containers in two networks: an external network and an internal network.
OC-Daemon runs as VPN client in the external network. Ocserv is
connected to both networks and enables a connected VPN client to reach
the internal network. Each network contains one web server that can be
used for connectivity checks and trusted network detection. The tests
check OC-Daemon with VPN connections in various settings.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>