Wait until VPN is disconnected during OC-Client reconnect
The OC-Client reconnect command sends a disconnect request to OC-Daemon,
waits for the current VPN connection to be terminated and then sends a
new connect request. Currently, the client is able to send the connect
request before the connection is completely shut down which causes the
reconnect command to fail. So, wait in OC-Client until OC-Daemon reaches
state "disconnected" and OpenConnect is not running any more before
connecting again. Additionally, set the "disconnected" and "OpenConnect
not running" states later in OC-Daemon.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>