Always allow DNS traffic with Traffic Policing
If a captive portal is detected, Traffic Policing sets the Allowed Ports
in the nftables rules to a list of captive portal ports (80, 443).
However, this removes the existing entry for the DNS port. So, set
static entries for DNS traffic in the nftables rules for Traffic
Policing and use the Allowed Ports only for captive portal ports.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>